django/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
click==8.1.7
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
phonenumbers==8.13.27
django-phonenumber-field==7.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
pyotp==2.9.0
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
twisted==23.10.0
txaio==23.1.1
twilio==8.11.0
typing-extensions==4.8.0
uvicorn==0.24.0.post1
zope.interface==6.1
